check.py: do not stringify DEP in stats.json
Let "dep" field in stats.json be a structure, not a string, so
that it is more easily parsed and it's view does not depend on
the dict key order. Maybe in future string id tag will also be
added (keys should be sorted to get stable DEP representation for
that, probably hashed afterwards).